requirements.txt
APScheduler==3.11.0
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pyTelegramBotAPI==4.14.1
python-dotenv==1.1.0
requests==2.32.3
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
pytz==2023.3